backend/voice_service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
librosa==0.10.1
scikit-learn==1.3.2
numpy==1.24.3